server/requirements.txt
google-generativeai==0.3.0
flask
flask-cors
#google-generativeai
python-dotenv
